review-feedback : 인수테스트시 실제 요청을 담당하는 클래스 추가 및 리팩토링
