+++ src/test/java/nextstep/subway/member/MemberAcceptanceTest.java
@@ -1,20 +1,27 @@
 package nextstep.subway.member;
 
 import static nextstep.subway.auth.acceptance.AuthAcceptanceTest.로그인_되어있음;
-import static org.assertj.core.api.Assertions.assertThat;
+import static nextstep.subway.behaviors.MemberBehaviors.내정보_삭제;
+import static nextstep.subway.behaviors.MemberBehaviors.내정보_삭제됨;
+import static nextstep.subway.behaviors.MemberBehaviors.내정보_수정;
+import static nextstep.subway.behaviors.MemberBehaviors.내정보_수정됨;
+import static nextstep.subway.behaviors.MemberBehaviors.내정보_조회;
+import static nextstep.subway.behaviors.MemberBehaviors.내정보_조회됨;
+import static nextstep.subway.behaviors.MemberBehaviors.회원_삭제_요청;
+import static nextstep.subway.behaviors.MemberBehaviors.회원_삭제됨;
+import static nextstep.subway.behaviors.MemberBehaviors.회원_생성됨;
+import static nextstep.subway.behaviors.MemberBehaviors.회원_생성을_요청;
+import static nextstep.subway.behaviors.MemberBehaviors.회원_정보_수정_요청;
+import static nextstep.subway.behaviors.MemberBehaviors.회원_정보_수정됨;
+import static nextstep.subway.behaviors.MemberBehaviors.회원_정보_조회_요청;
+import static nextstep.subway.behaviors.MemberBehaviors.회원_정보_조회됨;
 
-import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.AcceptanceTest;
 import nextstep.subway.auth.dto.TokenResponse;
-import nextstep.subway.member.dto.MemberRequest;
-import nextstep.subway.member.dto.MemberResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
 public class MemberAcceptanceTest extends AcceptanceTest {
     public static final String EMAIL = "email@email.com";
@@ -23,71 +30,6 @@ public class MemberAcceptanceTest extends AcceptanceTest {
     public static final String NEW_PASSWORD = "newpassword";
     public static final int AGE = 20;
     public static final int NEW_AGE = 21;
-
-    public static ExtractableResponse<Response> 회원_생성을_요청(String email, String password, Integer age) {
-        MemberRequest memberRequest = new MemberRequest(email, password, age);
-
-        return RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(memberRequest)
-                .when().post("/members")
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> 회원_정보_조회_요청(ExtractableResponse<Response> response) {
-        String uri = response.header("Location");
-
-        return RestAssured
-                .given().log().all()
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().get(uri)
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> 회원_정보_수정_요청(ExtractableResponse<Response> response, String email,
-                                                            String password, Integer age) {
-        String uri = response.header("Location");
-        MemberRequest memberRequest = new MemberRequest(email, password, age);
-
-        return RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(memberRequest)
-                .when().put(uri)
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> 회원_삭제_요청(ExtractableResponse<Response> response) {
-        String uri = response.header("Location");
-        return RestAssured
-                .given().log().all()
-                .when().delete(uri)
-                .then().log().all()
-                .extract();
-    }
-
-    public static void 회원_생성됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-    }
-
-    public static void 회원_정보_조회됨(ExtractableResponse<Response> response, String email, int age) {
-        MemberResponse memberResponse = response.as(MemberResponse.class);
-        assertThat(memberResponse.getId()).isNotNull();
-        assertThat(memberResponse.getEmail()).isEqualTo(email);
-        assertThat(memberResponse.getAge()).isEqualTo(age);
-    }
-
-    public static void 회원_정보_수정됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    public static void 회원_삭제됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
-    }
 
     @DisplayName("회원 정보를 관리한다.")
     @Test
@@ -135,44 +77,5 @@ public class MemberAcceptanceTest extends AcceptanceTest {
         ExtractableResponse<Response> deleteResponse = 내정보_삭제(사용자토큰);
         //then
         내정보_삭제됨(deleteResponse);
-
-
-    }
-
-    public static ExtractableResponse<Response> 내정보_조회(String token) {
-            return RestAssured
-                    .given().log().all()
-                    .auth().oauth2(token)
-                    .when().get("/members/me")
-                    .then().log().all()
-                    .extract();
-        }
-    public static ExtractableResponse<Response> 내정보_수정(String token, String email, String password, int age) {
-        MemberRequest updateRequest = new MemberRequest(email,password,age);
-        return RestAssured
-                .given().log().all().contentType(ContentType.JSON)
-                .auth().oauth2(token)
-                .when().body(updateRequest).put("/members/me")
-                .then().log().all()
-                .extract();
-    }
-    public static ExtractableResponse<Response> 내정보_삭제(String token) {
-        return RestAssured
-                .given().log().all()
-                .auth().oauth2(token)
-                .when().delete("/members/me")
-                .then().log().all()
-                .extract();
-    }
-    public static void 내정보_조회됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    public static void 내정보_수정됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    public static void 내정보_삭제됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
     }
 }
