+++ src/test/java/nextstep/subway/path/PathAcceptanceTest.java
@@ -1,20 +1,17 @@
 package nextstep.subway.path;
 
+import static nextstep.subway.behaviors.SubwayBehaviors.도착역으로가는_최단경로를_조회한다;
+import static nextstep.subway.behaviors.SubwayBehaviors.지하철_노선_등록되어_있음;
+import static nextstep.subway.behaviors.SubwayBehaviors.지하철_노선에_지하철역_등록_요청;
+import static nextstep.subway.behaviors.SubwayBehaviors.지하철역_등록되어_있음;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import nextstep.subway.AcceptanceTest;
-import nextstep.subway.line.acceptance.LineAcceptanceTest;
-import nextstep.subway.line.acceptance.LineSectionAcceptanceTest;
 import nextstep.subway.line.dto.LineResponse;
 import nextstep.subway.path.dto.PathResponse;
-import nextstep.subway.station.StationAcceptanceTest;
 import nextstep.subway.station.dto.StationResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -32,19 +29,7 @@ public class PathAcceptanceTest extends AcceptanceTest {
     private StationResponse 교대역;
     private StationResponse 남부터미널역;
 
-    public static ExtractableResponse<Response> 도착역으로가는_최단경로를_조회한다(StationResponse source, StationResponse target) {
-        Map<String, String> queryParams = new HashMap<>();
-        queryParams.put("source", String.valueOf(source.getId()));
-        queryParams.put("target", String.valueOf(target.getId()));
 
-        return RestAssured
-                .given().log().all()
-                .accept(ContentType.JSON)
-                .queryParams(queryParams)
-                .when().get("/paths")
-                .then().log().all()
-                .extract();
-    }
 
     /**
      * 교대역    --- *2호선* ---   강남역 |                        | *3호선*                   *신분당선* |                        |
@@ -54,16 +39,16 @@ public class PathAcceptanceTest extends AcceptanceTest {
     public void setUp() {
         super.setUp();
 
-        강남역 = StationAcceptanceTest.지하철역_등록되어_있음("강남역").as(StationResponse.class);
-        양재역 = StationAcceptanceTest.지하철역_등록되어_있음("양재역").as(StationResponse.class);
-        교대역 = StationAcceptanceTest.지하철역_등록되어_있음("교대역").as(StationResponse.class);
-        남부터미널역 = StationAcceptanceTest.지하철역_등록되어_있음("남부터미널역").as(StationResponse.class);
+        강남역 = 지하철역_등록되어_있음("강남역").as(StationResponse.class);
+        양재역 = 지하철역_등록되어_있음("양재역").as(StationResponse.class);
+        교대역 = 지하철역_등록되어_있음("교대역").as(StationResponse.class);
+        남부터미널역 = 지하철역_등록되어_있음("남부터미널역").as(StationResponse.class);
 
-        신분당선 = LineAcceptanceTest.지하철_노선_등록되어_있음("신분당선", "bg-red-600", 강남역, 양재역, 10);
-        이호선 = LineAcceptanceTest.지하철_노선_등록되어_있음("이호선", "bg-red-600", 교대역, 강남역, 10);
-        삼호선 = LineAcceptanceTest.지하철_노선_등록되어_있음("삼호선", "bg-red-600", 교대역, 양재역, 5);
+        신분당선 = 지하철_노선_등록되어_있음("신분당선", "bg-red-600", 강남역, 양재역, 10);
+        이호선 = 지하철_노선_등록되어_있음("이호선", "bg-red-600", 교대역, 강남역, 10);
+        삼호선 = 지하철_노선_등록되어_있음("삼호선", "bg-red-600", 교대역, 양재역, 5);
 
-        LineSectionAcceptanceTest.지하철_노선에_지하철역_등록_요청(삼호선, 교대역, 남부터미널역, 3);
+        지하철_노선에_지하철역_등록_요청(삼호선, 교대역, 남부터미널역, 3);
     }
 
     @Test
@@ -71,7 +56,7 @@ public class PathAcceptanceTest extends AcceptanceTest {
         StationResponse 출발역 = 교대역;
         StationResponse 도착역 = 양재역;
 
-        ExtractableResponse<Response> response = 도착역으로가는_최단경로를_조회한다(교대역, 양재역);
+        ExtractableResponse<Response> response = 도착역으로가는_최단경로를_조회한다(출발역, 도착역);
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
 
         PathResponse pathResponse = response.as(PathResponse.class);
